Add Ip and Device traffic methods
